# MTLS_UI/src/openapi/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/assignments/{id}": {
            "get": {
                "operationId": "AssignmentController_findOne",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài tập",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về thông tin bài tập"
                    },
                    "404": {
                        "description": "Không tìm thấy bài tập"
                    }
                },
                "summary": "Lấy thông tin bài tập theo ID",
                "tags": ["assignments"]
            },
            "patch": {
                "operationId": "AssignmentController_update",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài tập",
                    "schema": {
                        "type": "string"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAssignmentDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bài tập đã được cập nhật"
                    },
                    "404": {
                        "description": "Không tìm thấy bài tập"
                    }
                },
                "summary": "Cập nhật bài tập",
                "tags": ["assignments"]
            },
            "delete": {
                "operationId": "AssignmentController_remove",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài tập",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Bài tập đã được xóa"
                    },
                    "404": {
                        "description": "Không tìm thấy bài tập"
                    }
                },
                "summary": "Xóa bài tập",
                "tags": ["assignments"]
            }
        },
...
        "/do-assignments": {
            "post": {
                "operationId": "DoAssignmentController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDoAssignmentDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bài đã được nộp thành công."
                    }
                },
                "security": [{
                    "bearer": []
                }],
                "summary": "Nộp bài assignment",
                "tags": ["do-assignments"]
            },
            "get": {
                "operationId": "DoAssignmentController_findAll",
                "parameters": [{
                        "name": "assignment",
                        "required": false,
                        "in": "query",
                        "description": "ID của assignment để lọc bài nộp",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "student",
                        "required": false,
                        "in": "query",
                        "description": "ID của học viên để lọc bài nộp",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "teacher",
                        "required": false,
                        "in": "query",
                        "description": "ID của giáo viên để lọc bài nộp",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ungraded",
                        "required": false,
                        "in": "query",
                        "description": "Lọc bài nộp chưa chấm điểm",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách bài nộp."
                    }
                },
                "summary": "Lấy danh sách tất cả bài nộp hoặc theo filter",
                "tags": ["do-assignments"]
            }
        },
        "/do-assignments/{id}": {
            "get": {
                "operationId": "DoAssignmentController_findOne",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài nộp",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Thông tin bài nộp."
                    },
                    "404": {
                        "description": "Bài nộp không tồn tại."
                    }
                },
                "summary": "Lấy thông tin một bài nộp theo ID",
                "tags": ["do-assignments"]
            },
            "patch": {
                "operationId": "DoAssignmentController_update",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài nộp",
                    "schema": {
                        "type": "string"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDoAssignmentDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bài nộp đã được cập nhật."
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ."
                    },
                    "404": {
                        "description": "Bài nộp không tồn tại."
                    }
                },
                "security": [{
                    "bearer": []
                }],
                "summary": "Cập nhật thông tin bài nộp (chấm điểm)",
                "tags": ["do-assignments"]
            },
            "delete": {
                "operationId": "DoAssignmentController_remove",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài nộp",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Bài nộp đã được xóa."
                    },
                    "404": {
                        "description": "Bài nộp không tồn tại."
                    }
                },
                "summary": "Xóa một bài nộp",
                "tags": ["do-assignments"]
            }
        },
        "/do-assignments/{id}/detailed": {
            "get": {
                "operationId": "DoAssignmentController_findOneDetailed",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài nộp",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Thông tin chi tiết bài nộp."
                    },
                    "404": {
                        "description": "Bài nộp không tồn tại."
                    }
                },
                "summary": "Lấy thông tin chi tiết một bài nộp theo ID",
                "tags": ["do-assignments"]
            }
        },
        "/do-assignments/student/{studentId}/assignment/{assignmentId}": {
            "get": {
                "operationId": "DoAssignmentController_findByStudentAndAssignment",
                "parameters": [{
                        "name": "studentId",
                        "required": true,
                        "in": "path",
                        "description": "ID của học viên",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "assignmentId",
                        "required": true,
                        "in": "path",
                        "description": "ID của assignment",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách bài nộp của học viên và assignment."
                    },
                    "404": {
                        "description": "Học viên hoặc assignment không tồn tại."
                    }
                },
                "summary": "Lấy danh sách bài nộp theo student ID và assignment ID",
                "tags": ["do-assignments"]
            }
        },
        "/do-assignments/assignment/{assignmentId}": {
            "get": {
                "operationId": "DoAssignmentController_findByAssignment",
                "parameters": [{
                    "name": "assignmentId",
                    "required": true,
                    "in": "path",
                    "description": "ID của assignment",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Danh sách bài nộp của assignment."
                    },
                    "404": {
                        "description": "Assignment không tồn tại."
                    }
                },
                "summary": "Lấy danh sách bài nộp theo assignment ID",
                "tags": ["do-assignments"]
            }
        },
        "/do-assignments/student/{studentId}": {
            "get": {
                "operationId": "DoAssignmentController_findByStudent",
                "parameters": [{
                    "name": "studentId",
                    "required": true,
                    "in": "path",
                    "description": "ID của học viên",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Danh sách bài nộp của học viên."
                    },
                    "404": {
                        "description": "Học viên không tồn tại."
                    }
                },
                "summary": "Lấy danh sách bài nộp theo student ID",
                "tags": ["do-assignments"]
            }
        },
        "/do-assignments/teacher/{teacherId}": {
            "get": {
                "operationId": "DoAssignmentController_findByTeacher",
                "parameters": [{
                    "name": "teacherId",
                    "required": true,
                    "in": "path",
                    "description": "ID của giáo viên",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Danh sách bài nộp được chấm bởi giáo viên."
                    },
                    "404": {
                        "description": "Giáo viên không tồn tại."
                    }
                },
                "summary": "Lấy danh sách bài nộp theo teacher ID",
                "tags": ["do-assignments"]
            }
        },
        "/do-assignments/details/{id}": {
            "get": {
                "operationId": "DoAssignmentController_getSubmissionDetails",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài nộp",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Chi tiết bài nộp."
                    },
                    "404": {
                        "description": "Bài nộp không tồn tại."
                    }
                },
                "summary": "Lấy chi tiết đầy đủ về một bài nộp",
                "tags": ["do-assignments"]
            }
        },
...
        "/do-assignments/statistics/assignment/{assignmentId}": {
            "get": {
                "operationId": "DoAssignmentController_getAssignmentStatistics",
                "parameters": [{
                    "name": "assignmentId",
                    "required": true,
                    "in": "path",
                    "description": "ID của assignment",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Thống kê điểm số."
                    },
                    "404": {
                        "description": "Assignment không tồn tại."
                    }
                },
                "summary": "Lấy thống kê điểm số cho một assignment",
                "tags": ["do-assignments"]
            }
        },
        "/do-assignments/students/assignment/{assignmentId}": {
            "get": {
                "operationId": "DoAssignmentController_getStudentSubmissionsByAssignment",
                "parameters": [{
                    "name": "assignmentId",
                    "required": true,
                    "in": "path",
                    "description": "ID của assignment",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Danh sách học sinh và trạng thái chấm điểm."
                    },
                    "404": {
                        "description": "Assignment không tồn tại."
                    }
                },
                "summary": "Lấy danh sách học sinh đã làm assignment và trạng thái chấm điểm",
                "tags": ["do-assignments"]
            }
        },
        "/do-assignments/assignments/student/{studentId}": {
            "get": {
                "operationId": "DoAssignmentController_getAssignmentSubmissionsByStudent",
                "parameters": [{
                    "name": "studentId",
                    "required": true,
                    "in": "path",
                    "description": "ID của học sinh",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Danh sách assignment và trạng thái chấm điểm."
                    },
                    "404": {
                        "description": "Học sinh không tồn tại."
                    }
                },
                "summary": "Lấy danh sách assignment mà học sinh đã làm và trạng thái chấm điểm",
                "tags": ["do-assignments"]
            }
        },
        "/do-assignments/graded/teacher/{teacherId}": {
            "get": {
                "operationId": "DoAssignmentController_getGradedSubmissionsByTeacher",
                "parameters": [{
                    "name": "teacherId",
                    "required": true,
                    "in": "path",
                    "description": "ID của giáo viên",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Danh sách bài đã chấm."
                    },
                    "404": {
                        "description": "Giáo viên không tồn tại."
                    }
                },
                "summary": "Lấy danh sách bài đã chấm của giáo viên",
                "tags": ["do-assignments"]
            }
        },
        "/do-assignments/{id}/student-update": {
            "patch": {
                "operationId": "DoAssignmentController_updateStudentSubmission",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài nộp",
                    "schema": {
                        "type": "string"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStudentSubmissionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bài nộp đã được cập nhật thành công"
                    },
                    "401": {
                        "description": "Không có quyền cập nhật bài nộp này"
                    },
                    "404": {
                        "description": "Không tìm thấy bài nộp"
                    }
                },
                "security": [{
                    "bearer": []
                }],
                "summary": "Học viên cập nhật bài nộp của mình",
                "tags": ["do-assignments"]
            }
        },
...
    },
    "info": {
        "title": "API Documentation",
        "description": "The API's MTLS description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "UpdateAssignmentDto": {
                "type": "object",
                "properties": {
                    "section": {
                        "type": "string",
                        "description": "ID của section mà assignment thuộc về",
                        "example": "6507d6f6c31c7948c831a123"
                    },
                    "title": {
                        "type": "string",
                        "description": "Tiêu đề của assignment",
                        "example": "Bài tập về nhà: Lập trình hướng đối tượng (Cập nhật)"
                    },
                    "questionText": {
                        "type": "string",
                        "description": "Nội dung câu hỏi của assignment",
                        "example": "Hãy viết một chương trình minh họa tính kế thừa và đa hình trong OOP"
                    },
                    "description": {
                        "type": "string",
                        "description": "Mô tả chi tiết về assignment",
                        "example": "Bài tập này giúp sinh viên hiểu rõ hơn về tính kế thừa và đa hình trong lập trình hướng đối tượng"
                    },
                    "fileUrl": {
                        "type": "string",
                        "description": "URL của file đính kèm (nếu có)",
                        "example": "https://example.com/files/assignment1_updated.pdf"
                    },
                    "duration": {
                        "type": "number",
                        "description": "Thời gian làm bài (tính bằng phút)",
                        "example": 90
                    }
                }
            },
...
        }
    }
}